clang-tidy

+++ src/common/point.hpp
@@ -6,11 +6,11 @@ namespace bblp::advent_of_code_2022 {
 class Point {
   public:
     constexpr Point() : x(0), y(0) {}
-    constexpr Point(const int32_t x, const int32_t y) : x(x), y(y) {}
+    constexpr Point(const int64_t x, const int64_t y) : x(x), y(y) {}
 
     [[nodiscard]] bool operator==(const Point& other) const noexcept { return (x == other.x) && (y == other.y); }
 
-    int32_t x;
-    int32_t y;
+    int64_t x;
+    int64_t y;
 };
 }  // namespace bblp::advent_of_code_2022
